Add common class CustomLineType
